requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
djangorestframework==3.16.1
idna==3.11
psycopg2-binary==2.9.11
python-decouple==3.8
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
